Move vizjson callback back to Loader file
Since used there for jsonp, and needs access to the Loader object
