--- src-browserify/core/loader/vizjson.js
@@ -1,13 +1,0 @@
-var Loader = require('../loader');
-
-module.exports = function(data) {
-  Loader.current && Loader.current(data);
-  // remove script
-  Loader.head.removeChild(Loader._script);
-  Loader._script = null;
-  // next element
-  var a = Loader.queue.shift();
-  if (a) {
-    Loader.get(a[0], a[1]);
-  }
-}
